fix bug

+++ snake.py
@@ -27,7 +27,7 @@ while key != 27:                                                   # While Esc k
     win.border(0)
     win.addstr(0, 2, 'Score : ' + str(score) + ' ')                # Printing 'Score' and
     win.addstr(0, 27, ' SNAKE ')                                   # 'SNAKE' strings
-    win.timeout(150 - (len(snake)/5 + len(snake)/10)%120)          # Increases the speed of Snake as its length increases
+    win.timeout(150 - (int(len(snake)/5 + len(snake)/10)%120))          # Increases the speed of Snake as its length increases
     
     prevKey = key                                                  # Previous key pressed
     event = win.getch()
